Set auth plugin to have a high priority.
When used in conjunction with the browser plugin, the browser plugin
would reply before auth got a chance to process, and thus no auth
would ever be applied and the directory would be immediately
browseable. Setting the priority on the listener guarantees auth gets
a higher priority than browser (which uses normal priority).